+++ src/check.rs
@@ -12,7 +12,7 @@ use crate::{
 pub fn check_command<W: std::io::Write>(
     stderr: &mut W,
     filenames: Vec<PathBuf>,
-    max_width: usize,
+    max_width: u32,
 ) -> anyhow::Result<()> {
     let mut diagnostics = Vec::new();
 
@@ -41,7 +41,7 @@ pub fn check_command<W: std::io::Write>(
 
 pub(crate) fn check_one_file(
     filename: Option<&str>,
-    max_width: usize,
+    max_width: u32,
     content: String,
 ) -> Result<Vec<Diagnostic>, anyhow::Error> {
     let breaker = LineBreaker::builder().max_width(max_width).build()?;
@@ -55,10 +55,10 @@ pub(crate) fn check_one_file(
         };
 
         let (precedings, _) = line.split_at(line_break);
-        let column_no = precedings.encode_utf16().fold(0, |acc, _| acc + 1);
+        let column_no = precedings.encode_utf16().fold(0u32, |acc, _| acc + 1);
         let diagnostic = Diagnostic::new(
             filename,
-            line_no,
+            line_no as u32,
             column_no,
             "W001".to_string(),
             format!("Line length exceeds {max_width} characters"),
